==> notebooks/experimental/cell2/1-generate-experimental-features_cell2.ipynb <==
# 1) Generate features -  experiments - cell1_211011_3436 (cell2)

import json
import matplotlib.pyplot as plt

import bluepyopt as bpopt
import bluepyopt.ephys as ephys

import MEAutility as mu
from pprint import pprint
import time
import numpy as np
import neuroplotlib as nplt

import multimodalfitting as mf

import sys
import shutil

from pathlib import Path
import os
from datetime import datetime

%matplotlib notebook

save_fig = False

base_folder = Path("../../..")
fig_folder = Path("../../figures")

cell_name = "cell1_211006_3148"  # "cell1_211006_3148" | "cell1_211011_3436"

data_folder = base_folder / "experimental_data" / cell_name

cell_models_folder = base_folder / "cell_models"
model_folder = cell_models_folder / cell_name
model_name = cell_name
fitting_folder = model_folder / "fitting"
extra_folder = data_folder / "extracellular"

## 1) Generate features and protocols with BluePyEfe

from bluepyefe.extract import read_recordings, extract_efeatures_at_targets, compute_rheobase,\
    group_efeatures, create_feature_protocol_files, convert_legacy_targets
from bluepyefe.plotting import plot_all_recordings_efeatures

with (extra_folder / "probe.json").open() as f:
    probe_info = json.load(f)
    probe_info["pitch"] = 17.5
    probe = mu.return_mea(info=probe_info)

eap = np.load(extra_folder / "template.npy")
locations = np.load(extra_folder / "locations.npy")
fs = 20000

ax_eap = mu.plot_mea_recording(eap, probe)

efeatures_output_directory = fitting_folder / "efeatures"
ephys_dir = data_folder / "patch_data"

from bluepyefe.extract import read_recordings, extract_efeatures_at_targets, compute_rheobase,\
    group_efeatures, create_feature_protocol_files, convert_legacy_targets
from bluepyefe.plotting import plot_all_recordings_efeatures

from multimodalfitting.efeatures_extraction import build_wcp_metadata, wcp_reader, get_ecode_targets

# select files for different repetitions

# define timings for this experiment
ecodes_cell_timings = {
    "IDthres": {
        'ton': 400,
        'toff': 670
    },
    "firepattern": {
        'ton': 500,
        'toff': 4100
    },
    "IV": {
        'ton': 400,
        'toff': 3400
    },
    "IDrest": {
        'ton': 400,
        'toff': 1750
    },
    "APWaveform": {
        'ton': 350,
        'toff': 400
    },
    "HyperDepol": {
        'ton': 400,
        'toff': 1120,
        'tmid': 850
    },
    "sAHP": {
        'ton': 400,
        'toff': 1325,
        'tmid': 650,
        'tmid2': 875
    },
    "PosCheops": {
        'ton': 1000,
        't1': 9000,
        't2': 10500,
        't3': 14500,
        't4': 16000,
        'toff': 18660
    }
}

if cell_name == "cell1_211006_3148":
    runs = [1, 2, 3, 4, 5]  # run1 --> different rheobase
elif cell_name == "cell1_211011_3436":
    runs = [3, 4, 5, 6]
    
ecode_names = list(ecodes_cell_timings.keys())

files_list = []

for run in runs:
    rep_dict = {}
    for ecode in ecode_names:
        for patch_file in ephys_dir.iterdir():
            if f"run{run}" in patch_file.name and ecode.lower() in patch_file.name:
                rep_dict[ecode] = patch_file
    files_list.append(rep_dict)

files_metadata = build_wcp_metadata(cell_id=cell_name, 
                                    files_list=files_list, 
                                    ecode_timings=ecodes_cell_timings, 
                                    repetition_as_different_cells=False)
pprint(files_metadata[cell_name])

cells = read_recordings(
    files_metadata=files_metadata,
    recording_reader=wcp_reader
)

# define target features for different protocols
targets = get_ecode_targets(ecodes_cell_timings)

targets

global_tolerance = 30
for target in targets:
    target["tolerance"] = global_tolerance

pprint(targets)

t_start = time.time()
extract_efeatures_at_targets(
    cells, 
    targets,
    efel_settings={'interp_step': 0.1,
                   'strict_stiminterval': True}
)
t_stop = time.time()
print(f"Elapsed time {t_stop - t_start}")

if cell_name == "cell1_211006_3148":
    majority = 0.4
else:
    majority = 0.2

compute_rheobase(
    cells,
    protocols_rheobase=['IDthres'],
    rheobase_strategy="majority",
    rheobase_settings={"majority": majority}
)

print(f"Cell rheobase: {cells[0].rheobase} nA")

protocols = group_efeatures(cells, targets, use_global_rheobase=True)

efeatures, protocol_definitions, current = create_feature_protocol_files(
    cells,
    protocols,
    output_directory=efeatures_output_directory,
    threshold_nvalue_save=1,
    write_files=True,
)

efeatures["IDrest_300"]["soma"]

## 2) Convert to BPO format and append extra features


from multimodalfitting.efeatures_extraction import compute_extra_features, convert_to_bpo_format,\
    append_extrafeatures_to_json

protocols_of_interest = ["IDrest_150", "IDrest_250", "IDrest_300", "IV_-100", "IV_-20", "APWaveform_290"]

# remove sag features from IV_-20
exclude_features = {"IV_-20": ['sag_amplitude', 'sag_ratio1', 'sag_ratio2']}
# if cell_name == "cell1_211006_3148":
#     exclude_features.update({"IDrest_150": ['burst_number'],
#                              "IDrest_250": ['burst_number'],
#                              "IDrest_300": ['burst_number']})

in_protocol_path = efeatures_output_directory / "protocols.json"
in_efeatures_path = efeatures_output_directory / "features.json"

## Threshold EAP and extract featrues

thresh_uV = 5

amp_eap = np.ptp(eap, 1)
above_tr = np.where(amp_eap > thresh_uV)
eap_above = eap[above_tr]
probe_info = probe.info
locations_above = locations[above_tr]
probe_info["pos"] = locations_above.tolist()
probe_above = mu.return_mea(info=probe_info)

fig_t, ax_t = plt.subplots(figsize=(10, 10))
ax_t = mu.plot_mea_recording(eap, probe, vscale=150, ax=ax_t)

min_chan_above, _ = np.unravel_index(np.argmin(eap_above), eap_above.shape)

second_chan_loc = np.array([-85, 50])
distances_sec = [np.linalg.norm(second_chan_loc - loc) for loc in locations_above[:, :2]]
second_chan_idx = np.argmin(distances_sec)

fig_max, axs = plt.subplots(nrows=2, sharex=True) 
axs[0].plot(eap_above[min_chan_above, 15:80], color="C0", lw=3)
axs[1].plot(eap_above[second_chan_idx, 15:80], color="C1", lw=3)
axs[0].axis("off")
axs[1].axis("off")

fig_t_max, ax_t = plt.subplots(figsize=(10, 10))
ax_t = mu.plot_mea_recording(eap_above, probe_above, vscale=150, ax=ax_t)
ax_t = mu.plot_mea_recording(eap_above, probe_above, channels=[min_chan_above],  colors="C0",
                             vscale=150, ax=ax_t, lw=2)
ax_t = mu.plot_mea_recording(eap_above, probe_above, channels=[second_chan_idx],  colors="C1",
                             vscale=150, ax=ax_t, lw=2)

if save_fig:
    fig_t.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/template_above.pdf", dpi=300, transparent=True)
    fig_max.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/template_max.pdf", dpi=300, transparent=True)
    fig_t_max.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/template_above_max.pdf", dpi=300, transparent=True)

# save thresholded probe_above.json and template
json.dump(probe_above.info, (efeatures_output_directory / "probe_BPO.json").open("w"))
np.save(efeatures_output_directory / "template_BPO.npy", eap_above)

extra_features = compute_extra_features(eap_above, fs, upsample=10)

pprint(extra_features.keys())

fig_p2t = mf.plot_feature_map_w_colorbar(extra_features["peak_trough_ratio"], probe_above, 
                                         feature_name="Peak to trough ratio", label="a.u.", 
                                         cmap="coolwarm")
fig_negdiff = mf.plot_feature_map_w_colorbar(extra_features["neg_peak_diff"] * 1000, probe_above, 
                                             feature_name="Negative peak time difference", label="ms",
                                             cmap="PRGn")

if save_fig:
    fig_p2t.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/p2t_ratio.pdf", transparent=True)
    fig_negdiff.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/neg_diff.pdf", transparent=True)
    fig_p2t.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/p2t_ratio.png", dpi=300, transparent=True)
    fig_negdiff.savefig("/Users/abuccino/Documents/Submissions/papers/multimodal/neg_diff.png", dpi=300, transparent=True)

### 3.1) "soma" features

in_efeatures = json.load(open(in_efeatures_path, 'r'))

out_protocol_path_soma = efeatures_output_directory / "protocols_BPO_soma.json"
out_efeatures_path_soma = efeatures_output_directory / "features_BPO_soma.json"

protocols_dict_soma, efeatures_dict_soma = convert_to_bpo_format(in_protocol_path, in_efeatures_path, 
                                                                 out_protocol_path_soma, 
                                                                 out_efeatures_path_soma, 
                                                                 protocols_of_interest=protocols_of_interest, 
                                                                 exclude_features=exclude_features,
                                                                 std_from_mean=None)

### 3.2) "all" strategy

out_protocol_path_all = efeatures_output_directory / "protocols_BPO_all.json"
out_efeatures_path_all = efeatures_output_directory / "features_BPO_all.json"

protocols_dict_all, efeatures_dict_all = convert_to_bpo_format(in_protocol_path, in_efeatures_path, 
                                                               out_protocol_path_all, 
                                                               out_efeatures_path_all, 
                                                               protocols_of_interest=protocols_of_interest, 
                                                               exclude_features=exclude_features,
                                                               std_from_mean=None)

channel_ids_all = None

extrafeatures_dict_all = append_extrafeatures_to_json(extra_features, protocol_name="IDrest_300",
                                                      efeatures_dict=efeatures_dict_all,
                                                      efeatures_path=out_efeatures_path_all, 
                                                      channel_ids=None,
                                                      single_channel_features=False, 
                                                      std_from_mean=None)

pprint(extrafeatures_dict_all["IDrest_300"]["MEA"])

### 3.3) "single" strategy

cell = mf.create_experimental_model(model_name=model_name,
                                    release=True)
sim = ephys.simulators.LFPySimulator(cell, cvode_active=True, electrode=probe_above, 
                                     mechanisms_directory=model_folder)

out_protocol_path_single = efeatures_output_directory / "protocols_BPO_single.json"
out_efeatures_path_single = efeatures_output_directory / "features_BPO_single.json"

protocols_dict_single, efeatures_dict_single = convert_to_bpo_format(in_protocol_path, in_efeatures_path, 
                                                                     out_protocol_path_single, 
                                                                     out_efeatures_path_single, 
                                                                     protocols_of_interest=protocols_of_interest,
                                                                     exclude_features=exclude_features,
                                                                     std_from_mean=None)

# ax_eap = mu.plot_mea_recording(eap_above / np.max(np.abs(eap_above), axis=1, keepdims=True), probe_above)

# channel_ids_single = mf.select_single_channels(cell, sim, probe_above)
channel_ids_single = [208, 21, 1, 188, 90, 211, 180]

channel_ids_single

std_from_mean = 0.05

extrafeatures_dict_single = append_extrafeatures_to_json(extra_features, protocol_name="IDrest_300",
                                                         efeatures_dict=efeatures_dict_single,
                                                         efeatures_path=out_efeatures_path_single, 
                                                         channel_ids=channel_ids_single,
                                                         single_channel_features=True, 
                                                         std_from_mean=std_from_mean)

### 3.4) "sections" strategy

out_protocol_path_sections = efeatures_output_directory / "protocols_BPO_sections.json"
out_efeatures_path_sections = efeatures_output_directory / "features_BPO_sections.json"

protocols_dict_sections, efeatures_dict_sections = convert_to_bpo_format(in_protocol_path, in_efeatures_path, 
                                                                         out_protocol_path_sections, 
                                                                         out_efeatures_path_sections, 
                                                                         protocols_of_interest=protocols_of_interest,
                                                                         exclude_features=exclude_features,
                                                                         std_from_mean=None)

# ax_eap = mu.plot_mea_recording(eap_above / np.max(np.abs(eap_above), axis=1, keepdims=True), probe_above)

# channel_ids_sections = mf.select_mea_sections(cell, sim, probe_above)
channel_ids_sections = [
    [14, 31,  41, 56, 58, 74, 134, 136, 143, 187, 206],
    [1, 4, 21, 28, 34, 43, 49, 55, 86, 92, 102, 113, 115, 116, 127, 147, 148, 158, 168, 184, 210, 220],
    [20, 22, 37, 42, 60, 84, 88, 90, 97, 99, 108, 110, 142, 163, 180, 182, 188, 189, 219],
    [10, 30, 45, 69, 125, 144, 156, 161, 164, 172, 177, 198, 211]
]

channel_ids_sections

extrafeatures_dict_sections = append_extrafeatures_to_json(extra_features, protocol_name="IDrest_300",
                                                           efeatures_dict=efeatures_dict_sections,
                                                           efeatures_path=out_efeatures_path_sections, 
                                                           channel_ids=channel_ids_sections,
                                                           single_channel_features=False, 
                                                           std_from_mean=None)

pprint(extrafeatures_dict_sections["IDrest_300"]["MEA"])

### Plot selected channels

from matplotlib.patches import Ellipse, Rectangle

fig_single, ax_single = plt.subplots()

for pos in probe.positions:
    rect = Rectangle(xy=pos[:2] - probe.info["size"], width=probe.info["size"][0],
                     height=probe.info["size"][1], edgecolor=f"orange",
                     fc=f"orange", lw=1, alpha=0.7)
    ax_single.add_patch(rect)


mf.plot_cell(cell, sim, detailed=False, exclude_sections=["soma"], ax=ax_single)
ellipse = Ellipse(xy=(0, 0), width=15, height=15, 
                  edgecolor="grey", color="grey", lw=2)
ax_single.add_patch(ellipse)

for i, chan in enumerate(channel_ids_single):
    rect = Rectangle(xy=probe_above.positions[chan, :2] - probe.info["size"], width=probe.info["size"][0],
                     height=probe.info["size"][1], edgecolor=f"C{i}",
                     fc=f"C{i}", lw=2, label=f"Channel {chan}", alpha=1)
    ax_single.add_patch(rect)
ax_single.legend()
ax_single.set_xlim(-300, 300)
ax_single.set_ylim(-300, 300)

fig_sections, ax_sections = plt.subplots()

for pos in probe.positions:
    rect = Rectangle(xy=pos[:2] - probe.info["size"], width=probe.info["size"][0],
                     height=probe.info["size"][1], edgecolor=f"orange",
                     fc=f"orange", lw=1, alpha=0.7)
    ax_sections.add_patch(rect)
    
mf.plot_cell(cell, sim, detailed=False, exclude_sections=["soma"], ax=ax_sections)
ellipse = Ellipse(xy=(0, 0), width=15, height=15, 
                  edgecolor="grey", color="grey", lw=2)
ax_sections.add_patch(ellipse)

for i, chan_sect in enumerate(channel_ids_sections):
    for c, chan in enumerate(chan_sect):
        if c == 0:
            label = f"Section {i}"
        else:
            label = None
        rect = Rectangle(xy=probe_above.positions[chan, :2] - probe.info["size"], width=probe.info["size"][0],
                         height=probe.info["size"][1], edgecolor=f"C{i}",
                         fc=f"C{i}", lw=1, label=label)
        ax_sections.add_patch(rect)
ax_sections.legend()
ax_sections.set_xlim(-300, 300)
ax_sections.set_ylim(-300, 300)

fig_single.savefig(fig_folder / f"figS1E_single_{cell_name}.pdf")
fig_sections.savefig(fig_folder / f"figS1F_sections_{cell_name}.pdf")
